feat: Number 토큰의 타입을 Double 로 변경

+++ src/main/kotlin/calculator/Token.kt
@@ -1,7 +1,7 @@
 package calculator
 
 sealed class Token {
-    data class Number(val value: Int) : Token()
+    data class Number(val value: Double) : Token()
 
     object Plus : Token()
